24 bugfix

+++ models_3/model_frame.py
@@ -18,11 +18,13 @@ else:
 
 ### Frame Parameters
 ### should have this saved with the dataset
-data_rootdir = "../data/pyramid_sets/mf_test"
+data_rootdir = "../data/pyramid_sets/box_pyramid"
 model_dir = "trained"
-batch_size = 32
+batch_size = 1000
 mode = "train" # {train, load, loadtrain}
 override_existing_dir = True
+run_on_folds = [1]
+print("running on folds", run_on_folds)
 
 modelname = sys.argv[1]
 if not override_existing_dir and os.path.exists(model_dir + "/" + modelname):
@@ -30,7 +32,8 @@ if not override_existing_dir and os.path.exists(model_dir + "/" + modelname):
     print("exiting...")
     sys.exit(0)
 elif override_existing_dir and os.path.exists(model_dir + "/" + modelname):
-    os.system("rm " + model_dir + "/" + modelname + "/*")
+    if mode != "loadtrain":
+        os.system("rm " + model_dir + "/" + modelname + "/*")
 
 n_epochs_default = 20
 if len(sys.argv) > 2:
@@ -43,12 +46,18 @@ make_vis = ["performance map"]
 
 ### Model Parameters
 ### model types: cascade1, flat1,
-modeltype = "flat1"
-verbosity = 0
+modeltype = "cascade2"
+verbosity = 2
 model_parameters = {"cascade1": {"training_loss": "mse",
                                  "monitor_loss": True},
                     "flat1": {"training_loss": "mse",
-                                 "monitor_loss": True}
+                                 "monitor_loss": True},
+                    "cascade2": {"training_loss": "mse",
+                                 "monitor_loss": True,
+                                 "variant": "d",
+                                 "singletask": None},
+                    "flat2": {"training_loss": "mse",
+                                 "monitor_loss": True},
                     }
 
 ### iterate over row; iterate over comma separation
@@ -95,7 +104,8 @@ tr_wrangler = data_wrangler(data_rootdir, n_layers, n_folds, layer_dims, batch_s
 ### val wrangler
 va_wrangler = data_wrangler(data_rootdir, n_layers, n_folds, layer_dims, batch_size, buffer_nodata, x_layers, y_layers)
 
-computed_metrics = {"mse_single": [[] for ii in range(len(y_layers))],
+computed_metrics = {"metafolds": run_on_folds,
+                    "mse_single": [[] for ii in range(len(y_layers))],
                     "mse_each": [[] for ii in range(len(y_layers))],
                     "mse_overall": [[] for ii in range(len(y_layers))],
                     "mae_single": [[] for ii in range(len(y_layers))],
@@ -104,7 +114,7 @@ computed_metrics = {"mse_single": [[] for ii in range(len(y_layers))],
                     "r2": [[] for ii in range(len(y_layers))]}
 
 ### for each fold...
-for fold_i in range(n_folds):
+for fold_i in run_on_folds:
     print("running fold", fold_i)
     ### build model
     model = mb.make_model(modeltype=modeltype)
@@ -118,28 +128,44 @@ for fold_i in range(n_folds):
         model.load()
     if mode == "train" or mode == "loadtrain":
         print("fitting", fold_i)
-        model.fit(tr_wrangler, va_wrangler, n_epochs=n_epochs)
+        model.base_model.fit(tr_wrangler, callbacks=model.callbacks, epochs=n_epochs,
+                             validation_data=va_wrangler, verbose=2, workers=10,
+                             use_multiprocessing=True)
+        #model.fit(tr_wrangler, va_wrangler, n_epochs=n_epochs)
         model.save()
         ys, hs = model.predict(va_wrangler)
 
         ### compute metrics
 
 
-        for yl in range(len(y_layers)):
-            #mse_s, mse_e, mse_o = ms.metric_mse(ys[yl], hs[yl], "geo", ["single", "each", "overall"])
-            #mae_s, mae_e, mae_o = ms.metric_mae(ys[yl], hs[yl], "geo", ["single", "each", "overall"])
+        if modeltype != "cascade2" or model.singletask is None:
+            for yl in range(len(y_layers)):
+                #mse_s, mse_e, mse_o = ms.metric_mse(ys[yl], hs[yl], "geo", ["single", "each", "overall"])
+                #mae_s, mae_e, mae_o = ms.metric_mae(ys[yl], hs[yl], "geo", ["single", "each", "overall"])
 
-            mse_s = ms.metric_mse(ys[yl], hs[yl], "geo", ["single"])
-            mae_s = ms.metric_mae(ys[yl], hs[yl], "geo", ["single"])
+                mse_s = ms.metric_mse(ys[yl], hs[yl], "geo", ["single"])
+                mae_s = ms.metric_mae(ys[yl], hs[yl], "geo", ["single"])
+
+                computed_metrics["mse_single"][yl].append(mse_s)
+                #computed_metrics["mse_each"][yl].append(mse_e)
+                #computed_metrics["mse_overall"][yl].append(mse_o)
+                computed_metrics["mae_single"][yl].append(mae_s)
+                #computed_metrics["mae_each"][yl].append(mae_e)
+                #computed_metrics["mae_overall"][yl].append(mae_o)
+
+                computed_metrics["r2"][yl].append(ms.metric_r2(ys[yl], hs[yl]))
+        else:
+            yl = model.singletask
+            mse_s = ms.metric_mse(ys[0], hs[0], "geo", ["single"])
+            mae_s = ms.metric_mae(ys[0], hs[0], "geo", ["single"])
 
             computed_metrics["mse_single"][yl].append(mse_s)
-            #computed_metrics["mse_each"][yl].append(mse_e)
-            #computed_metrics["mse_overall"][yl].append(mse_o)
+            # computed_metrics["mse_each"][yl].append(mse_e)
+            # computed_metrics["mse_overall"][yl].append(mse_o)
             computed_metrics["mae_single"][yl].append(mae_s)
-            #computed_metrics["mae_each"][yl].append(mae_e)
-            #computed_metrics["mae_overall"][yl].append(mae_o)
+            # computed_metrics["mae_each"][yl].append(mae_e)
+            # computed_metrics["mae_overall"][yl].append(mae_o)
 
-            computed_metrics["r2"][yl].append(ms.metric_r2(ys[yl], hs[yl]))
-
+            computed_metrics["r2"][yl].append(ms.metric_r2(ys[0], hs[0]))
 mb.save_metrics(computed_metrics, model_dir + "/" + modelname, modelname)
